+++ src/main/java/pl/asie/computronics/Computronics.java
@@ -40,9 +40,9 @@ import pl.asie.computronics.block.BlockTapeReader;
 import pl.asie.computronics.cc.IntegrationComputerCraft;
 import pl.asie.computronics.cc.multiperipheral.MultiPeripheralRegistry;
 import pl.asie.computronics.integration.ModRecipes;
-import pl.asie.computronics.integration.buildcraft.ActionProvider;
-import pl.asie.computronics.integration.buildcraft.StatementParameters;
-import pl.asie.computronics.integration.buildcraft.TriggerProvider;
+import pl.asie.computronics.integration.buildcraft.statements.ActionProvider;
+import pl.asie.computronics.integration.buildcraft.statements.StatementParameters;
+import pl.asie.computronics.integration.buildcraft.statements.TriggerProvider;
 import pl.asie.computronics.integration.gregtech.GregTechRecipes;
 import pl.asie.computronics.integration.railcraft.IntegrationRailcraft;
 import pl.asie.computronics.item.ItemTape;
@@ -74,7 +74,7 @@ import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-@Mod(modid = Mods.Computronics, name = Mods.Computronics_NAME, version = "@VERSION@", useMetadata = true, dependencies = "required-after:asielib@[0.3.5,);after:ComputerCraft;after:OpenComputers@[1.4.0,);after:OpenComputers|Core;after:MineFactoryReloaded;after:RedLogic;after:ProjRed|Core;after:nedocomputers;after:BuildCraft|Core@[6.1.5,);after:Railcraft@[9.4.0.0,);after:gregtech;after:EnderIO;before:OpenPeripheralCore@[0.5.0,);before:OpenPeripheralApi@[2.2,)")
+@Mod(modid = Mods.Computronics, name = Mods.Computronics_NAME, version = "@VERSION@", useMetadata = true, dependencies = "required-after:asielib@[0.3.5,);after:ComputerCraft;after:OpenComputers@[1.4.0,);after:OpenComputers|Core;after:MineFactoryReloaded;after:RedLogic;after:ProjRed|Core;after:nedocomputers;after:BuildCraft|Core@[6.2.0,);after:Railcraft@[9.4.0.0,);after:gregtech;after:EnderIO;before:OpenPeripheralCore@[0.5.0,);before:OpenPeripheralApi@[2.2,)")
 public class Computronics {
 	public Config config;
 	public Compat compat;