+++ src/main/java/pl/asie/computronics/reference/Names.java
@@ -5,29 +5,31 @@ package pl.asie.computronics.reference;
  */
 public class Names {
 
-	public static final String AE2_SpatialIO = "spatial_io";
-	public static final String EnderIO_CapacitorBank = "capacitor_bank";
-	public static final String EnderIO_ExperienceTile = "experience_tile";
-	public static final String EnderIO_IOConfigurable = "io_config_tile";
-	public static final String EnderIO_MachineTile = "enderio_machine";
-	public static final String EnderIO_PoweredTile = "powered_tile";
-	public static final String EnderIO_RedstoneTile = "redstone_tile";
-	public static final String EnderIO_Transceiver = "dimensional_transceiver";
-	public static final String FSP_SteamTransporter = "steam_transporter";
-	public static final String FZ_ChargeConductor = "charge_conductor";
-	public static final String MFR_DSU = "deep_storage_unit";
-	public static final String Railcraft_BoilerFirebox = "boiler_firebox";
-	public static final String Railcraft_DigitalDetector = "digital_detector";
-	public static final String Railcraft_DigitalReceiverBox = "digital_receiver_box";
-	public static final String Railcraft_ElectricGrid = "electric_tile";
-	public static final String Railcraft_LauncherTrack = "launcher_track";
-	public static final String Railcraft_LimiterTrack = "limiter_track";
-	public static final String Railcraft_LocomotiveTrack = "locomotive_track";
-	public static final String Railcraft_PoweredTrack = "powered_track";
-	public static final String Railcraft_PrimingTrack = "priming_track";
-	public static final String Railcraft_RoutingDetector = "routing_detector";
-	public static final String Railcraft_RoutingSwitch = "routing_switch";
-	public static final String Railcraft_RoutingTrack = "routing_track";
-	public static final String Railcraft_SteamTurbine = "steam_turbine";
-	public static final String RedLogic_Lamp = "lamp";
+	public static final String
+		AE2_SpatialIO = "spatial_io",
+		BuildCraft_Heatable = "heatable_tile",
+		EnderIO_CapacitorBank = "capacitor_bank",
+		EnderIO_ExperienceTile = "experience_tile",
+		EnderIO_IOConfigurable = "io_config_tile",
+		EnderIO_MachineTile = "enderio_machine",
+		EnderIO_PoweredTile = "powered_tile",
+		EnderIO_RedstoneTile = "redstone_tile",
+		EnderIO_Transceiver = "dimensional_transceiver",
+		FSP_SteamTransporter = "steam_transporter",
+		FZ_ChargeConductor = "charge_conductor",
+		MFR_DSU = "deep_storage_unit",
+		Railcraft_BoilerFirebox = "boiler_firebox",
+		Railcraft_DigitalDetector = "digital_detector",
+		Railcraft_DigitalReceiverBox = "digital_receiver_box",
+		Railcraft_ElectricGrid = "electric_tile",
+		Railcraft_LauncherTrack = "launcher_track",
+		Railcraft_LimiterTrack = "limiter_track",
+		Railcraft_LocomotiveTrack = "locomotive_track",
+		Railcraft_PoweredTrack = "powered_track",
+		Railcraft_PrimingTrack = "priming_track",
+		Railcraft_RoutingDetector = "routing_detector",
+		Railcraft_RoutingSwitch = "routing_switch",
+		Railcraft_RoutingTrack = "routing_track",
+		Railcraft_SteamTurbine = "steam_turbine",
+		RedLogic_Lamp = "lamp";
 }